Separate out functions from request.ts
