+++ packages/client-library-otel/src/utils/request/index.ts
@@ -1,6 +1,4 @@
-export {
-  getRootRequestAttributes,
-  getRequestAttributes,
-  getResponseAttributes,
-  cloneRequestForAttributes,
-} from "./request";
+export { getIncomingRequestAttributes } from "./incoming";
+export { cloneRequestForAttributes } from "./clone";
+export { getResponseAttributes } from "./response";
+export { getRequestAttributes } from "./request";
